+++ CHANGES.md
@@ -1,6 +1,10 @@
 
 [back](https://github.com/MikeHardIce/strigui)
 
+### 0.0.1-alpha20
+
+* fixed an issue where a tabbed widget couldn't be tabbed again when removed and then readded
+
 ### 0.0.1-alpha19 
 
 * added skip-redraw? for widget updates